#19 Support Linux - Made the project compile and run on Ubuntu. Profiling doesn't work yet.

+++ ui/src/connect_dialog.cpp
@@ -90,7 +90,7 @@ void connect_dialog::accept()
         settings.setArrayIndex(index++);
         settings.setValue("ip", s.ip);
         settings.setValue("port", s.port);
-        settings.setValue("lastTime", s.time);
+        settings.setValue("lastTime", (qulonglong)s.time);
     }
     settings.endArray();
